Day 2 Part 1 Complete
After struggling to get the input to run successfully I’ve spotted my mistake and   have successfully reproduced the 1202 alarm.

The mistake was the way I was parsing the input, I was looking at runes.
By looking at runes instead of splitting on commas every digit was counted as a different input, so `10` was counted as `1`, `0`.